+++ system-parachains/asset-hubs/asset-hub-kusama/src/weights/mod.rs
@@ -44,8 +44,7 @@ pub mod pallet_proxy;
 pub mod pallet_recovery;
 pub mod pallet_remote_proxy;
 pub mod pallet_society;
-// TODO(#840): uncomment this so that pallet-revive is also benchmarked with this runtime
-// pub mod pallet_revive;
+pub mod pallet_revive;
 pub mod inmemorydb_weights;
 pub mod pallet_ah_ops;
 pub mod pallet_bags_list;